Add files via upload
This is my very first project based on forecasting.How much camping gear will one store sell each month in a year? To the uninitiated, calculating sales at this level may seem as difficult as predicting the weather. Both types of forecasting rely on science and historical data. In this competition, the fifth iteration, hierarchical sales data from Walmart will be used, the world’s largest company by revenue, to forecast daily sales for the next 28 days. It has explanatory variables such as price, promotions, day of the week, and special events. My work is to forecast of the unit sales of various products sold in the USA by Walmart.
It is a Kaggle project and it did not give the good result but someone can get the idea of coding of forecasting.